Extract MirrorSourceSyncer from MirrorMetadataManager
Split MirrorMetadataManager into two components with distinct
responsibilities and threading models:

MirrorMetadataManager keeps the MetadataPublisher implementation,
handling KRaft metadata updates and state transitions on the publisher
thread. It owns shared state (partition cache, source leader cache,
admin clients) and coordinator communication.

MirrorSourceSyncer handles all periodic sync from source clusters on
the scheduler thread: topic metadata, configs, group offsets, ACLs,
pattern discovery, epoch bumping, LME lookups, and mirror loop detection.

The syncer is created during initialize() and accesses shared state
through the manager. External callers (ClusterMirrorCoordinator,
MirrorFetcherThread, BrokerServer) are unchanged as thin delegation
methods on the manager forward to the syncer.
